pyproject.toml
[build-system]
requires = ["hatchling"]  # Moderner Build-Tool
build-backend = "hatchling.build"

[project]
name = "textswap"       # pip-name
version = "0.1.0"              # Deine Version
description = "Finds and Replace a String in Files, with Logger,Configreader and simple CLI"
readme = "README.md"
license = {text = "MIT"}       # Oder AGPL-3.0 etc.
authors = [
  {name = "Dein Name", email = "dein@email.de"}
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

requires-python = ">=3.8"
dependencies = [
  "pyyaml>=6.0",  # Dein ConfigReader
]

[project.optional-dependencies]
dev = [
]

[project.scripts]  # CLI-Befehle später
main = "textswap.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/textswap"]

# Linting/Formatting
[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "B", "Q"]
ignore = []

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
